mv QpileupFileReader, PileupFileReader and QsnpRecord back to qio
