--- qio/src/org/qcmg/qio/gff/GFFIterator.java
@@ -1,22 +1,0 @@
-package org.qcmg.gff;
-
-import java.io.InputStream;
-
-import org.qcmg.unused.record.AbstractRecordIterator;
-
-public class GFFIterator extends AbstractRecordIterator {
-
-	public GFFIterator(InputStream stream) throws Exception {
-		super(stream);
-	}
-
-	@Override
-	protected void readNext() {
-//		try {
-			next = GFFSerializer.nextRecord(reader);
-//		} catch (Exception ex) {
-//			next = null;
-//		}
-	}
-
-}
